+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `filament-simple-pages` will be documented in this file.
 
+## 0.4.0
+- option to publish Filament resource
+- removed setIcon(), setSort(), setNavigationGroup(), setNavigationLabel() (instead publish resource)
+- improved form layout with helper text
+- removed some useless dirs and files
+- add install command to make setup easier
+- improved readme/docs
+
 ## 0.3.0
 - add image feature by [@jetwes](https://github.com/jetwes)
 - add ->setIcon() option by [@jetwes](https://github.com/jetwes)